clean up

+++ hello.py
@@ -255,21 +255,30 @@ def update(id):
 
 #Delete record from databse 
 @app.route('/delete/<int:id>', methods=['GET','POST','DELETE'])
+@login_required
 def delete(id):
     user_to_delete = Users.query.get_or_404(id)
     name = None
     form = UserFrom()
-    try: 
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash('User Deleted Successfully!')
-        our_users=Users.query.order_by(Users.date_added)
-        return render_template('add_user.html', 
-                           form=form,
-                           name=name,
-                           our_users=our_users)
-    except:
-        flash('Error Occured!')
+    if current_user.id == id:
+        try: 
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash('User Deleted Successfully!')
+            our_users=Users.query.order_by(Users.date_added)
+            return render_template('add_user.html', 
+                               form=form,
+                               name=name,
+                               our_users=our_users)
+        except:
+            flash('Error Occured!')
+            our_users=Users.query.order_by(Users.date_added)
+            return render_template('add_user.html', 
+                               form=form,
+                               name=name,
+                               our_users=our_users)
+    else:
+        flash('Sorry you can\'t delete this user!')
         our_users=Users.query.order_by(Users.date_added)
         return render_template('add_user.html', 
                            form=form,
